create a file to categories tests
